ファイル名の変更を追加。選択ファイル数で表示するコンテキストメニューを変更。無選択の場合は削除確認を行わない。ツリーのコンテキストメニューを廃止。ツリーからの削除を廃止。

+++ src/main/scala/ice/DxFiler.scala
@@ -2,7 +2,7 @@ package ice
 
 import java.awt.datatransfer.{DataFlavor, Transferable, UnsupportedFlavorException}
 import java.awt.dnd.DnDConstants
-import java.awt.event.{InputEvent, KeyAdapter, KeyEvent, KeyListener, MouseAdapter, MouseEvent}
+import java.awt.event.*
 import java.awt.{Color, Desktop, Component as AComponent}
 import java.io.*
 import java.net.{URI, URL}
@@ -96,46 +96,81 @@ class DxFiler {
 
   private val table: FileTable = new FileTable
 
-  private def tablePopupMenu: PopupMenu = new PopupMenu {
-    private def createMenuItem(label: String, mnemonic: Key.Value)(f: => Unit): MenuItem = {
-      val action = Action(label) {
-        executeAndShowError {
-          f
+  private def showTablePopupMenu(invoker: Component, x: Int, y: Int): Unit = {
+    new PopupMenu {
+      private def createMenuItem(label: String, mnemonic: Key.Value)(f: => Unit): MenuItem = {
+        val action = Action(label) {
+          executeAndShowError {
+            f
+          }
+        }
+
+        val menuItem = new MenuItem(action)
+        menuItem.mnemonic = mnemonic
+        menuItem
+      }
+
+      if (getSelectedFiles.nonEmpty) {
+        if (desktop.isSupported(Desktop.Action.OPEN)) {
+          contents +=
+            createMenuItem("Open", Key.O) {
+              getSelectedFiles.foreach { file =>
+                executeAndShowError {
+                  openFile(file)
+                }
+              }
+            }
         }
       }
 
-      val menuItem = new MenuItem(action)
-      menuItem.mnemonic = mnemonic
-      menuItem
-    }
-
-    if (desktop.isSupported(Desktop.Action.OPEN)) {
-      contents +=
-        createMenuItem("Open", Key.O) {
-          getSelectedFiles.foreach { file =>
+      if (getSelectedFiles.length == 1) {
+        contents +=
+          createMenuItem("Rename", Key.R) {
+            val file = getSelectedFiles.head
             executeAndShowError {
-              openFile(file)
+              renameFile(file)
             }
           }
-        }
-    }
-
-    contents +=
-      createMenuItem("Delete", Key.D) {
-        getSelectedFiles.foreach { file =>
-          executeAndShowError {
-            deleteFile(file)
-          }
-        }
       }
+
+      if (getSelectedFiles.nonEmpty) {
+        contents +=
+          createMenuItem("Delete", Key.D) {
+            getSelectedFiles.foreach { file =>
+              executeAndShowError {
+                deleteFile(file)
+              }
+            }
+          }
+      }
+    }.show(invoker, x, y)
   }
 
   private def openFile(file: DxFile): Unit =
     WaitCursorWorker(frame, true)(() => desktop.open(Option(file.downloaded).getOrElse(Dx.download(file))))(null).execute()
 
+  private def renameFile(file: DxFile): Unit = {
+    if (file.toPath.toString == "/") throw new IllegalArgumentException("can't rename root path")
+
+    var toFile: DxFile = null
+
+    val worker =
+      WaitCursorWorker(frame, true) { () =>
+        Dialog.showInput(frame, "ファイル名", APP_NAME, initial = file.getName) match {
+          case Some(name: String) =>
+            toFile = file.toDxPath.parent.resolveDx(name).toDxFile
+            Dx.renameWithErrorMessage(file.toPath.toString, toFile.toPath.toString)
+            refresh()
+          case None =>
+        }
+      }(null)
+
+    worker.execute()
+  }
+
   private def deleteFile(file: DxFile): Unit =
     WaitCursorWorker(frame, true) { () =>
-      Dx.delete(file.toPath.toString)
+      Dx.deleteWithErrorMessage(file.toPath.toString)
       refresh()
     }(null).execute()
 
@@ -163,7 +198,7 @@ class DxFiler {
       ev.consume()
     case ev: MouseClicked
       if ev.clicks == 1 && (ev.modifiers & InputEvent.META_DOWN_MASK) != 0 =>
-      tablePopupMenu.show(table, ev.point.x, ev.point.y)
+      showTablePopupMenu(table, ev.point.x, ev.point.y)
     case ev: MouseClicked =>
       ev.consume()
     case ev: KeyPressed if ev.key == Key.Enter && ev.modifiers == 0 =>
@@ -190,21 +225,28 @@ class DxFiler {
       val rx = (r.x + r.width) min table.peer.getParent.getWidth
       val ry = (r.y + r.height) min table.peer.getParent.getHeight
 
-      tablePopupMenu.show(table, rx, ry)
+      showTablePopupMenu(table, rx, ry)
 
       ev.consume()
     case ev: KeyPressed if ev.key == Key.F5 =>
       refresh()
       ev.consume()
+    case ev: KeyPressed if ev.key == Key.F2 =>
+      if (getSelectedFiles.length == 1) {
+        renameFile(getSelectedFiles.head)
+      }
+      ev.consume()
     case ev: KeyPressed if ev.key == Key.Delete =>
-      Dialog.showConfirmation(frame, "削除しますか？", APP_NAME, Dialog.Options.OkCancel) match {
-        case Dialog.Result.Ok =>
-          getSelectedFiles.foreach { file =>
-            executeAndShowError {
-              deleteFile(file)
+      if (getSelectedFiles.nonEmpty ) {
+        Dialog.showConfirmation(frame, "削除しますか？", APP_NAME, Dialog.Options.OkCancel) match {
+          case Dialog.Result.Ok =>
+            getSelectedFiles.foreach { file =>
+              executeAndShowError {
+                deleteFile(file)
+              }
             }
-          }
-        case _ =>
+          case _ =>
+        }
       }
   }
 
@@ -235,52 +277,9 @@ class DxFiler {
     override def keyPressed(ev: KeyEvent): Unit = {
       if (ev.getKeyCode == KeyEvent.VK_F5) {
         refresh()
-      } else if (ev.getKeyCode == KeyEvent.VK_DELETE) {
-        Dialog.showConfirmation(frame, "削除しますか？", APP_NAME, Dialog.Options.OkCancel) match {
-          case Dialog.Result.Ok =>
-            val file = treePathToFile(tree.getSelectionPath)
-            WaitCursorWorker(frame, true) { () =>
-              Dx.delete(file.toPath.toString)
-              tree.setSelectionPath(tree.getSelectionPath.getParentPath)
-            }(null).execute()
-          case _ =>
-        }
-      } else if (ev.getKeyCode == KeyEvent.VK_CONTEXT_MENU) {
         ev.consume()
-        val r = tree.getPathBounds(tree.getSelectionPath)
-
-        val rx = (r.x + r.width) min table.peer.getParent.getWidth
-        val ry = (r.y + r.height) min table.peer.getParent.getHeight
-
-        showTreePopupMenu(tree, rx, ry)
       }
     }
-  }
-
-
-  def showTreePopupMenu(component: JComponent, x: Int, y: Int): Unit = {
-    new PopupMenu {
-      private def createMenuItem(label: String, mnemonic: Key.Value)(f: => Unit): MenuItem = {
-        val action = Action(label) {
-          executeAndShowError {
-            f
-          }
-        }
-
-        val menuItem = new MenuItem(action)
-        menuItem.mnemonic = mnemonic
-        menuItem
-      }
-
-      contents +=
-        createMenuItem("Delete", Key.D) {
-          val file = treePathToFile(tree.getSelectionPath)
-          WaitCursorWorker(frame, true) { () =>
-            Dx.delete(file.toPath.toString)
-            tree.setSelectionPath(tree.getSelectionPath.getParentPath)
-          }(null).execute()
-        }
-    }.show(Component.wrap(component), x, y)
   }
 
   private def getChildren(file: DxFile): Array[DxFile] =
@@ -320,19 +319,6 @@ class DxFiler {
     setCellRenderer(new FileTreeCellRenderer(getCellRenderer))
     expandRow(0)
     setVisibleRowCount(15)
-
-    addMouseListener(new MouseAdapter {
-      override def mouseClicked(ev: MouseEvent): Unit = {
-        if (ev.getClickCount == 1 && ev.getModifiersEx == InputEvent.META_DOWN_MASK) {
-          ev.consume()
-          val treePath = getPathForLocation(ev.getX, ev.getY)
-          if (treePath != null) {
-            if (getSelectionPaths == null || !getSelectionPaths.contains(treePath)) tree.setSelectionPath(treePath)
-            showTreePopupMenu(tree, ev.getX, ev.getY)
-          }
-        }
-      }
-    })
   }
 
   private def executeAndShowError(f: => Unit): Unit = {
